Update with mainline repo

+++ pr-check.sh
@@ -16,29 +16,29 @@
 
 # Check readiness for pull request
 
-set -ev
+set -e
 
 if [ ! -f brainiak/__init__.py ]
 then
-    echo "Run "$(basename "$0")" from the root of the BrainIAK repository"
+    echo "Run "$(basename "$0")" from the root of the BrainIAK hierarchy."
     exit 1
 fi
 
 basedir=$(pwd)
 
-function create_virtualenv_venv {
-    virtualenv ../$1
+function create_venv_venv {
+    python3 -m venv ../$1
 }
 
-function activate_virtualenv_venv {
+function activate_venv_venv {
     source ../$1/bin/activate
 }
 
-function deactivate_virtualenv_venv {
+function deactivate_venv_venv {
     deactivate
 }
 
-function remove_virtualenv_venv {
+function remove_venv_venv {
     rm -r ../$1
 }
 
@@ -48,9 +48,9 @@ function create_conda_venv {
 
 function activate_conda_venv {
     source activate $1
-    # Anaconda does not provide pip3 (at least on our Jenkins machine,
-    # Metacortex), so we install it in the virtual environment.
-    pip install -I pip
+    # Pip may update setuptools while installing BrainIAK requirements and
+    # break the Conda cached package, which breaks subsequent runs.
+    conda install --yes -f setuptools
 }
 
 function deactivate_conda_venv {
@@ -69,7 +69,6 @@ function exit_with_error {
 function exit_with_error_and_venv {
     $deactivate_venv
     cd $basedir
-    rm -f .coverage.*
     $remove_venv $venv
     exit_with_error "$1"
 }
@@ -82,49 +81,54 @@ then
     deactivate_venv=deactivate_conda_venv
     remove_venv=remove_conda_venv
     ignore_installed="--ignore-installed"
-elif [ $(which virtualenv) ]
-then
-    create_venv=create_virtualenv_venv
-    activate_venv=activate_virtualenv_venv
-    deactivate_venv=deactivate_virtualenv_venv
-    remove_venv=remove_virtualenv_venv
 else
-    echo "Cannot find virtualenv or conda."
-    echo "You must install one of them or test manually."
-    exit 1
+    create_venv=create_venv_venv
+    activate_venv=activate_venv_venv
+    deactivate_venv=deactivate_venv_venv
+    remove_venv=remove_venv_venv
 fi
 
-# optional, but highly recommended: create a virtualenv to isolate tests
+# Check if running in an sdist
+if git ls-files --error-unmatch pr-check.sh 2> /dev/null
+then
+    git clean -Xf .
+else
+    sdist_mode="--sdist-mode"
+fi
+
 venv=$(mktemp -u brainiak_pr_venv_XXXXX) || \
     exit_with_error "mktemp -u error"
 $create_venv $venv || {
-    exit_with_error "virtualenv creation failed"
+    exit_with_error "Virtual environment creation failed."
 }
 $activate_venv $venv || {
     $remove_venv $venv
-    exit_with_error "virtualenv activation failed"
+    exit_with_error "Virtual environment activation failed."
 }
 
 # install brainiak in editable mode (required for testing)
-pip3 install $ignore_installed -U -e . || \
-    exit_with_error_and_venv "pip3 failed to install BrainIAK"
+# brainiak will also be installed together with the developer dependencies, but
+# we install it first here to check that installation succeeds without the
+# developer dependencies.
+python3 -m pip install $ignore_installed -U -e . || \
+    exit_with_error_and_venv "Failed to install BrainIAK."
 
 # install developer dependencies
-pip3 install $ignore_installed -U -r requirements-dev.txt || \
-    exit_with_error_and_venv "pip3 failed to install requirements"
+python3 -m pip install $ignore_installed -U -r requirements-dev.txt || \
+    exit_with_error_and_venv "Failed to install development requirements."
 
 # static analysis
 ./run-checks.sh || \
     exit_with_error_and_venv "run-checks failed"
 
 # run tests
-./run-tests.sh || \
+./run-tests.sh $sdist_mode || \
     exit_with_error_and_venv "run-tests failed"
 
 # build documentation
 cd docs
-export THEANO_FLAGS='device=cpu,floatX=float64'
-git clean -Xf .
+export THEANO_FLAGS='device=cpu,floatX=float64,blas.ldflags=-lblas'
+
 if [ ! -z $SLURM_NODELIST ]
 then
     make_wrapper="srun -n 1"
@@ -135,8 +139,7 @@ $make_wrapper make || {
 }
 cd -
 
-# optional: remove virtualenv
 $deactivate_venv
 $remove_venv $venv
 
-echo "Check successful"
+echo "pr-check finished successfully."
